Test cleanup as prep for labware refactors
