+++ api/tests/opentrons/api/test_calibration.py
@@ -271,11 +271,8 @@ async def test_jog_calibrate_top(
     robot = model.robot
 
     container = model.container._container
-    pos1 = pose_tracker.change_base(
-        robot.poses,
-        src=container[0],
-        dst=robot.deck)
-    coordinates1 = container.coordinates()
+    pos1 = pose_tracker.absolute(robot.poses, container[0])
+    coordinates1 = container[0].coordinates()
 
     main_router.calibration_manager.move_to(model.instrument, model.container)
     main_router.calibration_manager.jog(model.instrument, 1, 'x')
@@ -289,7 +286,7 @@ async def test_jog_calibrate_top(
     )
 
     pos2 = pose_tracker.absolute(robot.poses, container[0])
-    coordinates2 = container.coordinates()
+    coordinates2 = container[0].coordinates()
 
     assert isclose(pos1 + (1, 2, 3), pos2).all()
     assert isclose(
